+++ src/features/member/api/oauth-login.ts
@@ -9,7 +9,7 @@ type LoginResponse = {
 export const Login = (provider: LowercaseOAuthType) => {
   return fetchRequest<LoginResponse>(ENDPOINTS.MEMBERS.AUTH.LOGIN(provider), {
     query: {
-      origin: encodeURIComponent(window.location.origin),
+      origin: window.location.origin,
     },
   })
 }
